requirements.txt
tensorflow
pillow
python-osc
numpy
pandas
sockets